+++ aws-java-sdk-rekognition/src/main/java/com/amazonaws/services/rekognition/AmazonRekognition.java
@@ -136,6 +136,12 @@ import com.amazonaws.services.rekognition.waiters.AmazonRekognitionWaiters;
  * <ul>
  * <li>
  * <p>
+ * <a
+ * href="https://docs.aws.amazon.com/rekognition/latest/APIReference/API_CopyProjectVersion.html">CopyProjectVersion</a>
+ * </p>
+ * </li>
+ * <li>
+ * <p>
  * <a href="https://docs.aws.amazon.com/rekognition/latest/APIReference/API_CreateDataset.html">CreateDataset</a>
  * </p>
  * </li>
@@ -158,6 +164,13 @@ import com.amazonaws.services.rekognition.waiters.AmazonRekognitionWaiters;
  * <li>
  * <p>
  * <a href="https://docs.aws.amazon.com/rekognition/latest/APIReference/API_DeleteProject.html">DeleteProject</a>
+ * </p>
+ * </li>
+ * <li>
+ * <p>
+ * <a
+ * href="https://docs.aws.amazon.com/rekognition/latest/APIReference/API_DeleteProjectPolicy.html">DeleteProjectPolicy
+ * </a>
  * </p>
  * </li>
  * <li>
@@ -204,6 +217,18 @@ import com.amazonaws.services.rekognition.waiters.AmazonRekognitionWaiters;
  * <p>
  * <a
  * href="https://docs.aws.amazon.com/rekognition/latest/APIReference/API_ListDatasetLabels.html">ListDatasetLabels</a>
+ * </p>
+ * </li>
+ * <li>
+ * <p>
+ * <a
+ * href="https://docs.aws.amazon.com/rekognition/latest/APIReference/API_ListProjectPolicies.html">ListProjectPolicies
+ * </a>
+ * </p>
+ * </li>
+ * <li>
+ * <p>
+ * <a href="https://docs.aws.amazon.com/rekognition/latest/APIReference/API_PutProjectPolicy.html">PutProjectPolicy</a>
  * </p>
  * </li>
  * <li>
@@ -367,6 +392,12 @@ import com.amazonaws.services.rekognition.waiters.AmazonRekognitionWaiters;
  * </a>
  * </p>
  * </li>
+ * <li>
+ * <p>
+ * <a href="https://docs.aws.amazon.com/rekognition/latest/APIReference/API_UpdateStreamProcessor.html">
+ * UpdateStreamProcessor</a>
+ * </p>
+ * </li>
  * </ul>
  */
 @Generated("com.amazonaws:aws-java-sdk-code-generator")
@@ -524,6 +555,66 @@ public interface AmazonRekognition {
      * @sample AmazonRekognition.CompareFaces
      */
     CompareFacesResult compareFaces(CompareFacesRequest compareFacesRequest);
+
+    /**
+     * <p>
+     * Copies a version of an Amazon Rekognition Custom Labels model from a source project to a destination project. The
+     * source and destination projects can be in different AWS accounts but must be in the same AWS Region. You can't
+     * copy a model to another AWS service.
+     * </p>
+     * <p>
+     * To copy a model version to a different AWS account, you need to create a resource-based policy known as a
+     * <i>project policy</i>. You attach the project policy to the source project by calling <a>PutProjectPolicy</a>.
+     * The project policy gives permission to copy the model version from a trusting AWS account to a trusted account.
+     * </p>
+     * <p>
+     * For more information creating and attaching a project policy, see Attaching a project policy (SDK) in the
+     * <i>Amazon Rekognition Custom Labels Developer Guide</i>.
+     * </p>
+     * <p>
+     * If you are copying a model version to a project in the same AWS account, you don't need to create a project
+     * policy.
+     * </p>
+     * <note>
+     * <p>
+     * To copy a model, the destination project, source project, and source model version must already exist.
+     * </p>
+     * </note>
+     * <p>
+     * Copying a model version takes a while to complete. To get the current status, call <a>DescribeProjectVersions</a>
+     * and check the value of <code>Status</code> in the <a>ProjectVersionDescription</a> object. The copy operation has
+     * finished when the value of <code>Status</code> is <code>COPYING_COMPLETED</code>.
+     * </p>
+     * 
+     * @param copyProjectVersionRequest
+     * @return Result of the CopyProjectVersion operation returned by the service.
+     * @throws AccessDeniedException
+     *         You are not authorized to perform the action.
+     * @throws InternalServerErrorException
+     *         Amazon Rekognition experienced a service issue. Try your call again.
+     * @throws InvalidParameterException
+     *         Input parameter violated a constraint. Validate your parameter before calling the API operation again.
+     * @throws LimitExceededException
+     *         An Amazon Rekognition service limit was exceeded. For example, if you start too many Amazon Rekognition
+     *         Video jobs concurrently, calls to start operations (<code>StartLabelDetection</code>, for example) will
+     *         raise a <code>LimitExceededException</code> exception (HTTP status code: 400) until the number of
+     *         concurrently running jobs is below the Amazon Rekognition service limit.
+     * @throws ResourceNotFoundException
+     *         The resource specified in the request cannot be found.
+     * @throws ThrottlingException
+     *         Amazon Rekognition is temporarily unable to process the request. Try your call again.
+     * @throws ServiceQuotaExceededException
+     *         <p>
+     *         The size of the collection exceeds the allowed limit. For more information, see Guidelines and quotas in
+     *         Amazon Rekognition in the Amazon Rekognition Developer Guide.
+     * @throws ProvisionedThroughputExceededException
+     *         The number of requests exceeded your throughput limit. If you want to increase this limit, contact Amazon
+     *         Rekognition.
+     * @throws ResourceInUseException
+     *         The specified resource is already being used.
+     * @sample AmazonRekognition.CopyProjectVersion
+     */
+    CopyProjectVersionResult copyProjectVersion(CopyProjectVersionRequest copyProjectVersionRequest);
 
     /**
      * <p>
@@ -912,7 +1003,8 @@ public interface AmazonRekognition {
      * </p>
      * <p>
      * <code>DeleteProject</code> is an asynchronous operation. To check if the project is deleted, call
-     * <a>DescribeProjects</a>. The project is deleted when the project no longer appears in the response.
+     * <a>DescribeProjects</a>. The project is deleted when the project no longer appears in the response. Be aware that
+     * deleting a given project will also delete any <code>ProjectPolicies</code> associated with that project.
      * </p>
      * <p>
      * This operation requires permissions to perform the <code>rekognition:DeleteProject</code> action.
@@ -938,6 +1030,36 @@ public interface AmazonRekognition {
      * @sample AmazonRekognition.DeleteProject
      */
     DeleteProjectResult deleteProject(DeleteProjectRequest deleteProjectRequest);
+
+    /**
+     * <p>
+     * Deletes an existing project policy.
+     * </p>
+     * <p>
+     * To get a list of project policies attached to a project, call <a>ListProjectPolicies</a>. To attach a project
+     * policy to a project, call <a>PutProjectPolicy</a>.
+     * </p>
+     * 
+     * @param deleteProjectPolicyRequest
+     * @return Result of the DeleteProjectPolicy operation returned by the service.
+     * @throws AccessDeniedException
+     *         You are not authorized to perform the action.
+     * @throws InternalServerErrorException
+     *         Amazon Rekognition experienced a service issue. Try your call again.
+     * @throws InvalidParameterException
+     *         Input parameter violated a constraint. Validate your parameter before calling the API operation again.
+     * @throws ResourceNotFoundException
+     *         The resource specified in the request cannot be found.
+     * @throws ThrottlingException
+     *         Amazon Rekognition is temporarily unable to process the request. Try your call again.
+     * @throws ProvisionedThroughputExceededException
+     *         The number of requests exceeded your throughput limit. If you want to increase this limit, contact Amazon
+     *         Rekognition.
+     * @throws InvalidPolicyRevisionIdException
+     *         The supplied revision id for the project policy is invalid.
+     * @sample AmazonRekognition.DeleteProjectPolicy
+     */
+    DeleteProjectPolicyResult deleteProjectPolicy(DeleteProjectPolicyRequest deleteProjectPolicyRequest);
 
     /**
      * <p>
@@ -2437,6 +2559,36 @@ public interface AmazonRekognition {
 
     /**
      * <p>
+     * Gets a list of the project policies attached to a project.
+     * </p>
+     * <p>
+     * To attach a project policy to a project, call <a>PutProjectPolicy</a>. To remove a project policy from a project,
+     * call <a>DeleteProjectPolicy</a>.
+     * </p>
+     * 
+     * @param listProjectPoliciesRequest
+     * @return Result of the ListProjectPolicies operation returned by the service.
+     * @throws AccessDeniedException
+     *         You are not authorized to perform the action.
+     * @throws InternalServerErrorException
+     *         Amazon Rekognition experienced a service issue. Try your call again.
+     * @throws InvalidParameterException
+     *         Input parameter violated a constraint. Validate your parameter before calling the API operation again.
+     * @throws ResourceNotFoundException
+     *         The resource specified in the request cannot be found.
+     * @throws ThrottlingException
+     *         Amazon Rekognition is temporarily unable to process the request. Try your call again.
+     * @throws ProvisionedThroughputExceededException
+     *         The number of requests exceeded your throughput limit. If you want to increase this limit, contact Amazon
+     *         Rekognition.
+     * @throws InvalidPaginationTokenException
+     *         Pagination token in the request is not valid.
+     * @sample AmazonRekognition.ListProjectPolicies
+     */
+    ListProjectPoliciesResult listProjectPolicies(ListProjectPoliciesRequest listProjectPoliciesRequest);
+
+    /**
+     * <p>
      * Gets a list of stream processors that you have created with <a>CreateStreamProcessor</a>.
      * </p>
      * 
@@ -2485,6 +2637,64 @@ public interface AmazonRekognition {
      * @sample AmazonRekognition.ListTagsForResource
      */
     ListTagsForResourceResult listTagsForResource(ListTagsForResourceRequest listTagsForResourceRequest);
+
+    /**
+     * <p>
+     * Attaches a project policy to a Amazon Rekognition Custom Labels project in a trusting AWS account. A project
+     * policy specifies that a trusted AWS account can copy a model version from a trusting AWS account to a project in
+     * the trusted AWS account. To copy a model version you use the <a>CopyProjectVersion</a> operation.
+     * </p>
+     * <p>
+     * For more information about the format of a project policy document, see Attaching a project policy (SDK) in the
+     * <i>Amazon Rekognition Custom Labels Developer Guide</i>.
+     * </p>
+     * <p>
+     * The response from <code>PutProjectPolicy</code> is a revision ID for the project policy. You can attach multiple
+     * project policies to a project. You can also update an existing project policy by specifying the policy revision
+     * ID of the existing policy.
+     * </p>
+     * <p>
+     * To remove a project policy from a project, call <a>DeleteProjectPolicy</a>. To get a list of project policies
+     * attached to a project, call <a>ListProjectPolicies</a>.
+     * </p>
+     * <p>
+     * You copy a model version by calling <a>CopyProjectVersion</a>.
+     * </p>
+     * 
+     * @param putProjectPolicyRequest
+     * @return Result of the PutProjectPolicy operation returned by the service.
+     * @throws AccessDeniedException
+     *         You are not authorized to perform the action.
+     * @throws InternalServerErrorException
+     *         Amazon Rekognition experienced a service issue. Try your call again.
+     * @throws InvalidParameterException
+     *         Input parameter violated a constraint. Validate your parameter before calling the API operation again.
+     * @throws InvalidPolicyRevisionIdException
+     *         The supplied revision id for the project policy is invalid.
+     * @throws MalformedPolicyDocumentException
+     *         The format of the project policy document that you supplied to <code>PutProjectPolicy</code> is
+     *         incorrect.
+     * @throws ResourceNotFoundException
+     *         The resource specified in the request cannot be found.
+     * @throws ResourceAlreadyExistsException
+     *         A resource with the specified ID already exists.
+     * @throws ThrottlingException
+     *         Amazon Rekognition is temporarily unable to process the request. Try your call again.
+     * @throws ServiceQuotaExceededException
+     *         <p>
+     *         The size of the collection exceeds the allowed limit. For more information, see Guidelines and quotas in
+     *         Amazon Rekognition in the Amazon Rekognition Developer Guide.
+     * @throws ProvisionedThroughputExceededException
+     *         The number of requests exceeded your throughput limit. If you want to increase this limit, contact Amazon
+     *         Rekognition.
+     * @throws LimitExceededException
+     *         An Amazon Rekognition service limit was exceeded. For example, if you start too many Amazon Rekognition
+     *         Video jobs concurrently, calls to start operations (<code>StartLabelDetection</code>, for example) will
+     *         raise a <code>LimitExceededException</code> exception (HTTP status code: 400) until the number of
+     *         concurrently running jobs is below the Amazon Rekognition service limit.
+     * @sample AmazonRekognition.PutProjectPolicy
+     */
+    PutProjectPolicyResult putProjectPolicy(PutProjectPolicyRequest putProjectPolicyRequest);
 
     /**
      * <p>