+++ aws-java-sdk-rekognition/src/main/java/com/amazonaws/services/rekognition/model/transform/ProjectVersionDescriptionJsonUnmarshaller.java
@@ -104,6 +104,10 @@ public class ProjectVersionDescriptionJsonUnmarshaller implements Unmarshaller<P
                     context.nextToken();
                     projectVersionDescription.setMaxInferenceUnits(context.getUnmarshaller(Integer.class).unmarshall(context));
                 }
+                if (context.testExpression("SourceProjectVersionArn", targetDepth)) {
+                    context.nextToken();
+                    projectVersionDescription.setSourceProjectVersionArn(context.getUnmarshaller(String.class).unmarshall(context));
+                }
             } else if (token == END_ARRAY || token == END_OBJECT) {
                 if (context.getLastParsedParentElement() == null || context.getLastParsedParentElement().equals(currentParentElement)) {
                     if (context.getCurrentDepth() <= originalDepth)